lambda/src/thumbnail-creation/thumbnail_creation/requirements.txt
boto3==1.14.38
pillow==7.2.0
nose